+++ examples/make-fuzzer-input.cpp
@@ -25,7 +25,7 @@
 
 #include <deepstream/buffer.hpp>
 #include <deepstream/message_builder.hpp>
-#include <deepstream/random.hpp>
+#include "../src/random.hpp"
 
 #include <cassert>
