+++ src/random.cpp
@@ -18,7 +18,7 @@
 
 #include <deepstream/buffer.hpp>
 #include <deepstream/message_builder.hpp>
-#include <deepstream/random.hpp>
+#include "random.hpp"
 
 #include <cassert>
